Add files via upload
This is a hopefully working version of the snapshot. There is no capability to connect the the front end currently. It will ask the user for glide ratio, plane velocity, and plane altitude. It will iterate the bank angle from 0 to 45 on increments of 5. It will absolute value these coordinates to get the mirror if the bank angle isn't 0.